test will now seed only once
