+++ server/test/03_review.spec.js
@@ -10,37 +10,39 @@ chai.use(chaiHttp);
 chai.use(chaiJSON);
 chai.should();
 
-const route = '/api/v1/ratings';
+const route = '/api/v1/reviews';
 
 async function getSample(valid = true) {
 
   if (valid) {
     const chapter = (await knex('chapters')
       .select(['chapters.id'])
-      .where({approved: true})
+      .where({ approved: true })
+      .limit(1))[0];
+
+    const rating = (await knex('ratings')
+      .select(['ratings.id', 'ratings.reaction'])
+      .where({ isDeleted: false })
       .limit(1))[0];
 
     return {
       chapterId: chapter.id,
-      reaction: 'like',
+      reaction: rating.reaction,
+      ratingId: rating.id,
       metadata: {
-        audioQuality: 1,
-        contentAccuracy: 1,
-        language: 2
+        audioQuality: 'It was good',
+        contentAccuracy: ['Zero accuracy'],
+        language: 'No profanity or cursing works'
       }
     };
 
   }
 }
 
-describe('CHAPTER RATINGS ROUTE', () => {
-  before(async () => {
-    await knex.migrate.rollback();
-    await knex.migrate.latest();
-    return await knex.seed.run();
-  });
+describe('CHAPTER REVIEWS ROUTE', () => {
 
-  it('Should fetch existing chapter ratings', done => {
+
+  it('Should fetch existing chapter reviews', done => {
     chai
       .request(server)
       .get(route)
@@ -48,67 +50,67 @@ describe('CHAPTER RATINGS ROUTE', () => {
       .end((err, res) => {
         res.should.have.status(200);
         res.should.be.json;
-        res.body.ratings[0].should.have.property('id');
+        res.body.reviews[0].should.have.property('id');
         done();
       });
   });
 
-  it('Should query existing chapter ratings to include respective review', done => {
+  it('Should query existing chapter reviews to include respective rating', done => {
     chai
       .request(server)
-      .get(`${route}?include=review`)
+      .get(`${route}?include=rating`)
       .set(tokens.headersSuperAdmin1)
       .end((err, res) => {
         res.should.have.status(200);
         res.should.be.json;
-        res.body.ratings[0].should.have.property('id');
-        res.body.ratings[0].should.have.property('review');
+        res.body.reviews[0].should.have.property('id');
+        res.body.reviews[0].should.have.property('rating');
         done();
       });
   });
 
-  it('Should create chapter ratings', async () => {
+  it('Should create chapter review', async () => {
     //clear everything to avoid conflict
-    await knex('ratings').delete();
+    await knex('reviews').delete();
     const data = await getSample();
     return chai
       .request(server)
       .post(route)
       .set(tokens.headersSuperAdmin1)
       .set('Content-Type', 'application/json')
-      .send({rating: data})
+      .send({ review: data })
       .then((res) => {
         res.status.should.eql(201);
         res.should.be.json;
-        res.body.should.have.property('rating');
+        res.body.should.have.property('review');
       });
   });
 
-  it('Should update a chapter rating', async () => {
-    const rating = (await knex('ratings').select().limit(1))[0];
+  it('Should update a chapter review', async () => {
+    const review = (await knex('reviews').select().limit(1))[0];
     const data = await getSample();
 
     return chai
       .request(server)
-      .put(`${route}/${rating.id}`)
+      .put(`${route}/${review.id}`)
       .set(tokens.headersSuperAdmin1)
       .set('Content-Type', 'application/json')
-      .send({rating: data})
+      .send({ review: data })
       .then((res) => {
         res.should.have.status(200);
         res.should.be.json;
-        res.body.rating.should.have.property('id');
-        res.body.rating.should.have.property('metadata');
-        res.body.rating.metadata.should.eql(data.metadata);
+        res.body.review.should.have.property('id');
+        res.body.review.should.have.property('metadata');
+        res.body.review.metadata.should.eql(data.metadata);
       });
   });
 
-  it('Should delete a chapter rating', async () => {
-    const rating = (await knex('ratings').select().limit(1))[0];
+  it('Should delete a chapter review', async () => {
+    const review = (await knex('reviews').select().limit(1))[0];
 
     return chai
       .request(server)
-      .delete(`${route}/${rating.id}`)
+      .delete(`${route}/${review.id}`)
       .set(tokens.headersSuperAdmin1)
       .then((res) => {
         res.status.should.eql(200);
