+++ server/test/06_h5p.spec.js
@@ -32,9 +32,6 @@ const editorModelSchema = {
 
 
 describe('H5P Route', () => {
-  before(async () => {
-  });
-
   it('Should return editor model object', done => {
     chai
       .request(server)
